backend/Django_xm/requirements.txt
aiofiles>=25.1.0
aiohappyeyeballs>=2.6.1
aiohttp>=3.13.3
aiosignal>=1.4.0
aiosqlite>=0.22.1
amqp>=5.3.1
annotated-doc>=0.0.4
annotated-types>=0.7.0
anyio>=4.13.0
asgiref>=3.11.1
asttokens>=3.0.1
attrs>=26.1.0
backcall>=0.2.0
beautifulsoup4>=4.14.3
billiard>=4.2.4
bleach>=6.3.0
blis>=1.3.3
catalogue>=2.0.10
celery>=5.4.0
celery-types>=0.26.0
certifi>=2026.2.25
cffi>=2.0.0
charset-normalizer>=3.4.6
click>=8.3.1
click-didyoumean>=0.3.1
click-plugins>=1.1.1.2
click-repl>=0.3.0
cloudpathlib>=0.23.0
colorama>=0.4.6
confection>=1.3.3
cryptography>=46.0.6
cymem>=2.0.13
dataclasses-json>=0.6.7
decorator>=5.2.1
deepagents>=0.1.0
defusedxml>=0.7.1
distro>=1.9.0
Django>=5.2.12
django-cors-headers>=4.9.0
django-filter>=25.2
django-redis>=6.0.0
django-simpleui>=2026.1.13
django-stubs>=6.0.1
django-stubs-ext>=6.0.1
djangorestframework>=3.17.1
djangorestframework-stubs>=3.16.8
djangorestframework_simplejwt>=5.5.1
dnspython>=2.8.0
docopt>=0.6.2
drf-spectacular>=0.29.0
emoji>=2.15.0
en_core_web_sm>=3.8.0
eventlet>=0.41.0
executing>=2.2.1
faiss-cpu>=1.13.2
fastjsonschema>=2.21.2
filelock>=3.25.2
filetype>=1.2.0
flower>=2.0.1
frozenlist>=1.8.0
greenlet>=3.3.2
h11>=0.16.0
html5lib>=1.1
httpcore>=1.0.9
httpx>=0.28.1
httpx-sse>=0.4.3
humanize>=4.15.0
idna>=3.11
inflection>=0.5.1
installer>=0.7.0
ipython>=8.12.3
jedi>=0.19.2
Jinja2>=3.1.6
jiter>=0.13.0
jsonpatch>=1.33
jsonpointer>=3.1.1
jsonschema>=4.26.0
jsonschema-specifications>=2025.9.1
jupyter_client>=8.8.0
jupyter_core>=5.9.1
jupyterlab_pygments>=0.3.0
kombu>=5.6.2
langchain>=1.2.13
langchain-classic>=1.0.3
langchain-community>=0.4.1
langchain-core>=1.2.22
langchain-openai>=1.0.2
langchain-anthropic>=0.4.1
langchain-groq>=1.1.0
langchain-chroma>=0.2.0
langchain-mcp-adapters>=0.1.0
langchain-milvus>=0.1.1
langchain-tavily>=0.2.17
langchain-text-splitters>=1.1.1
langdetect>=1.0.9
langgraph>=1.1.3
langgraph-checkpoint>=4.0.1
langgraph-checkpoint-sqlite>=3.0.3
langgraph-checkpoint-postgres>=2.0.0
langgraph-prebuilt>=1.0.8
langgraph-sdk>=0.3.12
langgraph-store-postgres>=0.1.0
langsmith>=0.7.22
llvmlite>=0.46.0
loguru>=0.7.3
lxml>=6.0.2
Markdown>=3.10.2
markdown-it-py>=4.0.0
MarkupSafe>=3.0.3
marshmallow>=3.26.2
matplotlib-inline>=0.2.1
mdurl>=0.1.2
mistune>=3.2.0
multidict>=6.7.1
murmurhash>=1.0.15
mypy_extensions>=1.1.0
mysqlclient>=2.2.8
nbclient>=0.10.4
nbconvert>=7.17.1
nbformat>=5.10.4
numba>=0.64.0
numpy>=2.4.3
olefile>=0.47
openai>=2.30.0
orjson>=3.11.7
ormsgpack>=1.12.2
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandocfilters>=1.5.1
parso>=0.8.6
pickleshare>=0.7.5
pillow>=12.1.1
pipreqs>=0.5.0
platformdirs>=4.9.6
preshed>=3.0.13
prometheus_client>=0.24.1
prompt_toolkit>=3.0.52
propcache>=0.4.1
psutil>=7.2.2
psycopg2-binary>=2.9.10
pure_eval>=0.2.3
pycparser>=3.0
pydantic>=2.12.5
pydantic-settings>=2.13.1
pydantic_core>=2.41.5
Pygments>=2.19.2
PyJWT>=2.12.1
PyMySQL>=1.1.2
pypdf>=6.9.2
pypdfium2>=5.6.0
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.1
python-iso639>=2026.1.31
python-magic>=0.4.27
python-multipart>=0.0.22
python-oxmsg>=0.0.2
pytz>=2026.1.post1
PyYAML>=6.0.3
pyzmq>=27.1.0
RapidFuzz>=3.14.3
redis>=5.3.1
referencing>=0.37.0
regex>=2026.2.28
requests>=2.33.0
requests-toolbelt>=1.0.0
rich>=14.3.3
rpds-py>=0.30.0
shellingham>=1.5.4
six>=1.17.0
smart_open>=7.5.1
sniffio>=1.3.1
socksio>=1.0.0
soupsieve>=2.8.3
spacy>=3.8.13
spacy-legacy>=3.0.12
spacy-loggers>=1.0.5
SQLAlchemy>=2.0.48
sqlite-vec>=0.1.7
sqlparse>=0.5.5
srsly>=2.5.3
stack-data>=0.6.3
tavily-python>=0.7.23
tenacity>=9.1.4
thinc>=8.3.13
tiktoken>=0.12.0
tinycss2>=1.4.0
tornado>=6.5.5
tqdm>=4.67.3
traitlets>=5.14.3
typer>=0.24.1
types-PyYAML>=6.0.12.20250915
typing-inspect>=0.9.0
typing-inspection>=0.4.2
typing_extensions>=4.15.0
tzdata>=2025.3
unstructured>=0.22.6
unstructured-client>=0.42.12
uritemplate>=4.2.0
urllib3>=2.6.3
uuid_utils>=0.14.1
vine>=5.1.0
wasabi>=1.1.3
wcwidth>=0.6.0
weasel>=1.0.0
webencodings>=0.5.1
win32_setctime>=1.2.0
wrapt>=2.1.2
xxhash>=3.6.0
yarg>=0.1.9
yarl>=1.23.0
zstandard>=0.25.0
